Require semicolons after break and continue, and make them do relative jumps

+++ src/program.hpp
@@ -67,12 +67,12 @@ struct op_jump_relative
 
 struct op_break
 {
-    std::size_t jump;
+    std::int64_t jump;
 };
 
 struct op_continue
 {
-    std::size_t jump;
+    std::int64_t jump;
 };
 
 struct op_jump_if_false
